fix(auth,frontend): split API-key store file, guard stale summary loads
Addresses the two still-open CodeRabbit findings on this PR.

store_postgres.go was past the project's 500-line ceiling and this PR
pushed it further. Its API-key surface moves to store_postgres_apikeys.go
(856 + 301 lines). Pure move: no query, scanning or signature changes.

loadApiKeysUsageStats could render a stale result. Concurrent refreshes
resolve in arbitrary order, so a slow earlier request could overwrite a
newer summary, or paint its error over a summary that had loaded fine. A
generation counter makes a load a no-op once a newer one has started, on
both the success and the failure path.

Adds apikeys-usage.test.ts, which the summary module had been missing
entirely: render path, top-keys list, markup-in-key-name staying literal
text, count-abbreviation boundaries, the error path, and both ordering
guards. The two ordering tests fail against the unguarded loader.
